requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
sgp4>=2.22
influxdb-client>=1.40.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
